fix(auto-routing-contracts): break index/normalize import cycle
normalize.ts imported types from index.ts while index.ts re-exports
normalize.ts, which failed apps/web's madge circular-dependency check.
The shared input schema now lives in input.ts and both import from it.
