+++ packages/auto-routing-contracts/src/index.ts
@@ -1,39 +1,11 @@
 import * as z from 'zod';
+import { NormalizedClassifierInputSchema } from './input';
 
-export type JsonValue =
-  | string
-  | number
-  | boolean
-  | null
-  | { [key: string]: JsonValue }
-  | JsonValue[];
-
-const JsonValueSchema: z.ZodType<JsonValue> = z.lazy(() =>
-  z.union([
-    z.string(),
-    z.number(),
-    z.boolean(),
-    z.null(),
-    z.array(JsonValueSchema),
-    z.record(z.string(), JsonValueSchema),
-  ])
-);
-
-export const NormalizedClassifierInputSchema = z.object({
-  apiKind: z.enum(['chat_completions', 'responses', 'messages']),
-  requestedModel: z.string(),
-  systemPromptPrefix: z.string().nullable(),
-  userPromptPrefix: z.string().nullable(),
-  latestUserPromptPrefix: z.string().nullable().optional(),
-  messageCount: z.number().int().nonnegative().nullable(),
-  hasTools: z.boolean(),
-  stream: z.boolean(),
-  providerHints: z.object({
-    provider: JsonValueSchema,
-    providerOptions: JsonValueSchema,
-  }),
-});
-export type NormalizedClassifierInput = z.infer<typeof NormalizedClassifierInputSchema>;
+export {
+  NormalizedClassifierInputSchema,
+  type JsonValue,
+  type NormalizedClassifierInput,
+} from './input';
 
 // What the gateway mirrors to the auto-routing worker per request: the
 // already-normalized classifier input plus caller identity. The gateway
